version definitive

+++ src/main/java/com/nawell/magasin/repositories/CommandeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CommandeRepository extends JpaRepository<Commande, Integer> {
+public interface CommandeRepository extends JpaRepository<Commande, Long> {
 }
